Got Spring boot to run; change Employee.dep type
Had to change the type of Employee.dep away from the pointer-ish Department and to a String defined as DEP_NAME in order for h2 compatibility. It was a more tiresome process to troubleshoot getting maven/spring to run properly, and tbh I'm not really sure of all the steps I took to do it.